workers: move landing_worker module (bug 1819145)
